Calculate the cloud cover of MODIS data within an aoi

calcMODIS_aoi_cloudcov estimates the cloud cover of Landsat data based on preview images using the haze-optimal transformation (HOT)

calcMODIS_aoi_cloudcov(records, aoi = NULL, maxDeviation = 20,
  cloudPrbThreshold = 40, slopeDefault = 1.4, interceptDefault = -10,
  dir_out = NULL, username = NULL, password = NULL, verbose = TRUE)

Arguments

records
data.frame, one or multiple records (each represented by one row), as it is returned by getSentinel_query.

aoi
sfc_POLYGON or SpatialPolygons or matrix, representing a single multi-point (at least three points) polygon of your area-of-interest (AOI). If it is a matrix, it has to have two columns (longitude and latitude) and at least three rows (each row representing one corner coordinate). If its projection is not +proj=longlat +datum=WGS84 +no_defs, it is reprojected to the latter. Use set_aoi instead to once define an AOI globally for all queries within the running session. If aoi is undefined, the AOI that has been set using set_aoi is used.

maxDeviation
numeric, the maximum allowed deviation of calculated scene cloud cover from the provided scene cloud cover. Use 100 if you do not like to consider the cloud cover % given by the data distributor. Default is maxDeviation = 20.

cloudPrbThreshold
numeric, the threshold of the HOT cloud probability layer (0-100 %) below which pixels are considered as clear sky. Default is cloudPrbThreshold = 40. It will be dynamically adjusted according to the input in maxDeviation.

slopeDefault
numeric, value taken as slope ONLY if least-alternate deviation regression fails. Default is 1.4, proven to work well for common land surfaces.f default values. In this case cloud cover will be set to 9999 % for the given record.

interceptDefault
numeric, value taken as intercept ONLY if least-alternate deviation regression fails. Default is -10, proven to work well for common land surfaces.

dir_out
character, optional. Full path to target directory where to save the cloud masks. If NULL, cloud masks are not saved.

username
character, a valid user name to the ESA Copernicus Open Access Hub. If NULL (default), the session-wide login credentials are used (see login_CopHub for details on registration).

password
character, the password to the specified user account. If NULL (default) and no seesion-wide password is defined, it is asked interactively ((see login_CopHub for details on registration).

verbose
logical, if TRUE, details on the function's progress will be visibile on the console. Default is TRUE.

Value

A data.frame as the records input with one additional column holding the estimated cloud cover within the aoi

Details

Cloud cover of optical remote sensing data is normally provided scene-wide by the data dissiminator. This function uses HOT in order to derive a local cloud cover % within an aoi. The estimation of the cloud cover is done on the red and blue information of the preview images provided by the respective data dissiminator. Haze-optimal transformation (HOT) procedure is applied based on Zhu & Helmer (2018), https://data.fs.usda.gov/research/pubs/iitf/ja_iitf_2018_Zhu.pdf. Orignally, the algorithm was introduced by Zhang et al. (2002) "An image transform to characterize and compensate for spatial variations in thin cloud contamination of Landsat images", Remote Sensing of Environment 82, 2-3. HOT seperates clear-sky pixels first from a threshold, calculates a least alternate deviation (LAD) regression from these pixels and exposes cloud pixels by the deviation of all pixels from this clear-sky line.

See also

Examples

# NOT RUN {
## Load packages
library(getSpatialData)
library(sf)

## set aoi and time range for the query
set_aoi(aoi_data[[1]])
time_range <-  c("2017-08-01", "2017-08-30")

## Login to USGS ERS
# }# NOT RUN {
login_USGS("username")

## set archive directory
set_archive("/path/to/archive/")

## get available products and select one
product_names <- getMODIS_names()
product <- grep("MOD13Q1", product_names, value = T)

## query for records for your AOI, time range and product
records <- getMODIS_query(time_range = time_range, name = product)

## Calculate cloud cover within the aoi
records_cloudcov <- calcMODIS_aoi_cloudcov(records = records, aoi = aoi) # cloud cov. calc.

## preview a record
getMODIS_preview(records_cloudcov[1,])

## download records 1 and 2
files <- getMODIS_data(records_cloudcov[1:2,])
# }# NOT RUN {
# }