Check download availability of records

check_availability checks for each record whether it is available for direct download (can be downloaded instantly) or not (and thus must be ordered before download).

check_availability(records, verbose = TRUE)

Arguments

records

records data frame, containing one or multiple records (each represented by one row), as returned by get_records

verbose

logical, whether to display details on the function's progress or output on the console.

Value

A data frame of records (as defined with argument records), extended by a column download_available (logical).

Note

To use this function, you must be logged in at the services required for your request. See the examples and login for details.

Author

Jakob Schwalb-Willmann