Extract Sentinel zip archives

extractSentinel unzips Sentinel zip archives downloaded using getSentinel_data.

extractSentinel(datasets, dir_out = NULL, junkpaths = T,
  list_only = F, pattern = NULL, verbose = T)

Arguments

datasets
character vector or list, containing paths to datasets (.zip, as downloaded by getSentinel_data) that should be extracted.

dir_out
character, full path to target directory. By default, extracted .SAFE folders are stored in the same directory as the zip originals (recommended).

junkpaths
logical, reduces the folder structure inside the extracted SAFE folder to one level. Although loosing the original folder structure, this is recommended especially on windows machine due to issues that can occure with very long paths.

list_only
logical, if TRUE, extracting is skipped and just the files contained in datasets are returned.

pattern
character, pattern (e.g. "xml") of files to be extracted.

verbose
logical, whether to display details on the function's progress or output on the console.

Value

List containing paths to the extracted .SAFE folders and all contained files.

See also