Preview a MOSIS image

getMODIS_preview previews a single MODIS record as image on a map or as RGB plot. The function is useful to apply visual checks to records before downloading them.

getMODIS_preview(record, on_map = TRUE, show_aoi = TRUE, verbose = TRUE)

Arguments

record
data.frame, single row data.frame collected from the return of getMODIS_query, representing the selected record and all its attributes.

on_map
logical, if TRUE, the preview is displaed corner-georeferenced on a map. If FALSE, a simple RGB plot is displayed. Default is TRUE.

show_aoi
logical, if TRUE, the session AOI defined with set_aoi is drawn to the map viewer. Ignored, if on_map = FALSE or if no AOI has been defined with set_aoi. Default is TRUE.

verbose
logical, if TRUE, details on the function's progress will be visibile on the console. Default is TRUE.

Value

None. A plot/view display is generated.

See also