Restore Sentinel datasets from Copernicus LTA

getSentinel_restore requests to restore Setninel datasets that have been archived by ESA to the Copernicus Long-Term Archive (LTA) (see argument check_avail of getSentinel_query).

getSentinel_restore(record, username = NULL, password = NULL,
  hub = "auto", verbose = TRUE)

Arguments

record
data.frame, single row data.frame collected from the return of getSentinel_query, representing the selected record and all its attributes.

username
character, a valid user name to the ESA Copernicus Open Access Hub. If NULL (default), the session-wide login credentials are used (see login_CopHub for details on registration).

password
character, the password to the specified user account. If NULL (default) and no seesion-wide password is defined, it is asked interactively ((see login_CopHub for details on registration).

hub
character, either

"auto" (default) to automatically select a suitable Copernicus hub depending on the selected platform,

"dhus" to look for operational Open Hub records only,

"s3" to look for Sentinel-3 pre-operational records only,

"s5p" to look for Sentinel-5 precursor pre-operational records only,

"GNSS" to look for GNSS RINEX records only,

or a valid API URL.

verbose
logical, if TRUE, details on the function's progress will be visibile on the console. Default is TRUE.

Value

TRUE on success.

See also