Get previews of records

get_previews downloads and georeferences preview images per records and saves them to disk. File paths are added to the records data frame.

get_previews(
  records,
  dir_out = NULL,
  force = FALSE,
  as_sf = TRUE,
  ...,
  verbose = TRUE
)

Arguments

records
records data frame, containing one or multiple records (each represented by one row), as returned by get_records

dir_out
character, a directory, to which the downloaded file(s) should be saved. By default, these are saved to the archive directory defined with set_archive.

force
logical, whether to force download of files already existing or not.

as_sf
logical, whether records should be returned as sf data.frame or a simple data.frame. In both cases, spatial geometries are stored in column footprint.

...
additional, sensor-specific arguments:

hub, character, Copernicus Hub selection for Sentinel queries. Either

"auto" (default) to automatically select a suitable Copernicus hub depending on the selected products

"dhus" to look for operational Open Hub records only,

"s5p" to look for Sentinel-5P precursor pre-operational records only,

"GNSS" to look for GNSS RINEX records only,

or a valid API URL.

verbose
logical, whether to display details on the function's progress or output on the console.

Value

A data frame of records (as defined with argument records), extended by two columns: preview_file (character, path to georeferenced preview) and preview_file_jpg (character, path to preview JPG).

Note

To use this function, you must be logged in at the services required for your request. See the examples and login for details.

Author

Jakob Schwalb-Willmann