Query available product records

get_records queries a service for available records using basic input search parameters such as product name (see get_products), AOI and time range. The function returns a data frame of records that can be further filtered and that other getSpatialData functions use as input.

get_records(
  time_range,
  products,
  aoi = NULL,
  as_sf = TRUE,
  rename_cols = TRUE,
  check_products = TRUE,
  simplify_cols = TRUE,
  ...,
  verbose = TRUE
)

getSentinel_records(
  time_range,
  products,
  aoi = NULL,
  as_sf = TRUE,
  rename_cols = TRUE,
  check_products = TRUE,
  simplify_cols = TRUE,
  ...,
  verbose = TRUE
)

getLandsat_records(
  time_range,
  products,
  aoi = NULL,
  as_sf = TRUE,
  rename_cols = TRUE,
  check_products = TRUE,
  simplify_cols = TRUE,
  ...,
  verbose = TRUE
)

getMODIS_records(
  time_range,
  products,
  aoi = NULL,
  as_sf = TRUE,
  rename_cols = TRUE,
  check_products = TRUE,
  simplify_cols = TRUE,
  ...,
  verbose = TRUE
)

getSRTM_records(
  time_range,
  products,
  aoi = NULL,
  as_sf = TRUE,
  rename_cols = TRUE,
  check_products = TRUE,
  simplify_cols = TRUE,
  ...,
  verbose = TRUE
)

Arguments

time_range

character, a vector of two elements: the query's starting date and stopping date, formatted "YYYY-MM-DD", e.g. c("2017-05-15", "2017-06-15")

products

character, product name(s). Use get_products to get a full list of all available products. If multiple products are supplied, the returned records are combined across products.

aoi

sfc_POLYGON or SpatialPolygons or matrix, representing a single multi-point (at least three points) polygon of your area-of-interest (AOI). If it is a matrix, it has to have two columns (longitude and latitude) and at least three rows (each row representing one corner coordinate). If its projection is not +proj=longlat +datum=WGS84 +no_defs, it is reprojected to the latter. Use set_aoi instead to once define an AOI globally for all queries within the running session. If aoi is undefined, the AOI that has been set using set_aoi is used.

as_sf

logical, whether records should be returned as sf data.frame or a simple data.frame. In both cases, spatial geometries are stored in column footprint.

rename_cols

logical, whether to rename columns to a product-independent standard to make it possible to combine records of different products received from different sources.

check_products

logical, check whether product names (products) are valid (recommended) or not. If FALSE, product names are not checked and directly used to attempt a query even if they are not supported or spelled wrong.

simplify_cols

logical, whether to simplify the returned records table by removing fields that are not standard across different products.

...

additional, sensor-specific arguments:

hub, character, Copernicus Hub selection for Sentinel queries. Either

"auto" (default) to automatically select a suitable Copernicus hub depending on the selected products

"dhus" to look for operational Open Hub records only,

"s5p" to look for Sentinel-5P precursor pre-operational records only,

"GNSS" to look for GNSS RINEX records only,

or a valid API URL.

verbose

logical, whether to display details on the function's progress or output on the console.

Value

A data frame of records (by default an sf data frame, see argument as_sf). Each row represents one record. The data frame can be further filtered by its columnwise attributes or plotted to view their spatial footprints. The records data frame can be used as input to other getSpatialData functions.

Details

To use these functions, you need to be logged in at the required services: To query Sentinel records, login with your ESA Copernicus Open Access Hub credentials using login_CopHub. To query MODIS and Landsat records, login with your USGS EROS Registration System (ERS) credentials using login_USGS. See login for details.

GNSS

GNSS products (such as "sentinel-1_gnss") retrieved from the dual-frequency GPS recievers mounted on Sentinel-1, -2, and -3 represent a special type of product, as they are AOI-independent and thus only referenced by mission time (argument time_range). GNSS data originally have been used only to precisely calculate the satellites' orbits, but then have been released to the scientific public due to their potential scientifc uses (for details, see https://earth.esa.int/web/sentinel/missions/sentinel-3/news/-/article/new-gnss-l1b-rinex-data-release-for-sentinel-1-2-and-3 and https://earth.esa.int/documents/247904/351187/GMES_Sentinels_POD_Service_File_Format_Specification).

Author

Jakob Schwalb-Willmann