Session-wide login

Functions to login to services supported by getSpatialData once for the running session and to recieve their current status.

login_CopHub(username = NULL, password = NULL, n_retry = 3, verbose = TRUE)

login_codede(username = NULL, password = NULL, n_retry = 3, verbose = TRUE)

login_USGS(username = NULL, password = NULL, n_retry = 3, verbose = TRUE)

login_earthdata(username = NULL, password = NULL, n_retry = 3, verbose = TRUE)

services(value = F, verbose = T)

Arguments

username
character, user name to login at the respective service.

password
character, password to login at the respective service.

n_retry
numeric, number of attempts to login, if login fails

verbose
logical, whether to display details on the function's progress or output on the console.

value
logical, whether to return a data frame containing service status or not. Default is FALSE.

Details

login_CopHub logs you in at the ESA Copernicus Open Access Hub using your credentials (register once at https://scihub.copernicus.eu/).

login_codede logs you in at CODE-DE using your credentials (register once at https://code-de.org/).

login_USGS logs you in at the USGS EROS Registration System (ERS) using your credentials (register once at https://ers.cr.usgs.gov/register/).

login_earthdata logs you in at the NASA Earth Data User Registration System (URS) using your credentials (register once at https://urs.earthdata.nasa.gov/users/new/)

services displays the status of all online services used by getSpatialData. Services that are operating as usual are labeled "available". Returns a data.frame containing the service status, if argument value is set to TRUE.

Note

Login credentials will be saved and made available for all getSpatialData functions during the whole session. They will be erased when quitting the session. Alternatively, login credentials can be set individually with each get* function call.

See also

Author

Jakob Schwalb-Willmann

Examples

library(getSpatialData)

if (FALSE) {
## Define user credentials for the Copernicus Open Access Hub
login_CopHub(username = "my_user_name", password = "my_password")

## Define USGS user credentials
login_USGS(username = "my_user_name", password = "my_password")

## get status of all services
services()
}