Read records from file

read_records reads records that have been written through write_records.

read_records(file, as_sf = TRUE, verbose = TRUE)

Arguments

file
character file path from where to read the records file.

as_sf
logical specifies if the records shall be returned as sf object. Otherwise a data.frame will be returned. Default is TRUE.

verbose
logical, if TRUE, details on the function's progress will be visibile on the console. Default is TRUE.

Value

sf object or data.frame depending on as_sf.

Author

Henrik Fisser, 2020