Overview of getSpatialData

Homogeneous and reproducible workflows to query, preview, analyse, select, order and download various kinds of spatial datasets from open sources.

Introduction

getSpatialData is an R package in an early development stage that ultimately aims to enable homogeneous and reproducible workflows to query, preview, analyse, select, order and download various kinds of spatial datasets from open sources.

Currently, getSpatialData supports a variety of products, including Sentinel-1, Sentinel-2, Sentinel-3, Sentinel-5P, Landsat 8 OLI, Landsat ETM, Landsat TM, Landsat MSS, MODIS (Terra & Aqua) and SRTM DEMs. For this, getSpatialData facilitates access to multiple services implementing clients to public APIs of ESA Copernicus Open Access Hub, USGS EarthExplorer, USGS EROS ESPA, Amazon Web Services (AWS), NASA DAAC LAADS and NASA CMR search. A full list of all supported products can be found below.

Workflow

The getSpatialData workflow to query, preview, analyse, select, order and download spatial data is designed to be as reproducible as possible and is made of six steps:

querying products of interest (see get_products) for available records by an area of interest (AOI) and time (see get_records),

previewing geometries and previews of the obtained records (see get_previews, view_records and view_previews),

analysing records by deriving scene and AOI cloud distribution and coverage directly from preview imagery (see calc_cloudcov),

selecting records based on user-defined measures (see select_unitemporal, select_bitemporal and select_timeseries),

ordering datasets that are not available for immediate download (on-demand) but need to be ordered or restored before download (see check_availability and order_data), and lastly

downloading the full datasets for those records that have been selected in the process, solely based on meta and preview-dervied data.

For all steps, getSpatialData supports local chaching to reduce bandwith usage and uneccasary downloads.

Further details are included in the README file, which is constantly updated on https://github.com/16EAGLE/getSpatialData.

See also

Author

Jakob Schwalb-Willmann

Henrik Fisser