Overview of getSpatialData tools

getSpatialData is an R package that aims to provide homogeneous function bundles to query, preview, download, prepare and transform various kinds of spatial datasets from open sources, e.g. Satellite sensor data and higher-level environmental data products.

Details

At the moment, the package includes the following functions:

Sentinel

getSentinel_query querys the Copernicus Open Access Hubs for Sentinel-1, -2 and -3 data and returns a data frame containing the found records (rows) and their attributes (columns).

getSentinel_preview uses the output of getSentinel_query to preview (quick-look) a user-selected record even before downloading it. By default, the preview is displayed corner-georeferenced in a map viewer in relation to the session AOI.

getSentinel_data uses the output of getSentinel_query to download Sentinel data.

Landsat

getLandsat_names obtains available Landsat product names from USGS Earth Explorer, which can be optionally used with getLandsat_query to narrow the search.

getLandsat_query querys USGS Earth Explorer for Landsat data and returns a data frame containing the found records (rows) and their attributes (columns).

getLandsat_preview uses the output of getLandsat_query to preview (quick-look) a user-selected record By default, the preview is displayed corner-georeferenced in a map viewer in relation to the session AOI.

getLandsat_data uses the output of getLandsat_query to order and download Landsat data.

supports order (on-demand processing) and download of higher-level products (all Landsat products), e.g. top-of-atmosphere (TOA), surface reflectance (SR) or different indices, from USGS-EROS ESPA.

supports direct download of Level-1 products (Landsat-8 only) via Amazon Web Services (AWS).

will support direct download of Level-1 products (all Landsat products) via USGS EarthExplorer (requires a USGS user profile with machine-to-machine download permission)

MODIS

getMODIS_names obtains available MODIS product names from USGS Earth Explorer, which can be optionally used with getMODIS_query to narrow the search.

getMODIS_query querys USGS Earth Explorer for MODIS data and returns a data frame containing the found records (rows) and their attributes (columns).

getMODIS_preview uses the output of getMODIS_query to preview (quick-look) a user-selected record. By default, the preview is displayed corner-georeferenced in a map viewer in relation to the session AOI.

getMODIS_data uses the output of getMODIS_query to order and download MODIS data from LAADS.

Login

login_CopHub define your Copernicus Open Access login credentials once for the present R session to be able to call each getSentinel function without defining login arguments each time you use them.

login_USGS define your USGS login credentials once for the present R session to be able to call each get* function that connects to a USGS service without defining login arguments each time you use them.

Session settings

set_archive define a getSpatialData archive directory to which all *_data functions will download data.

set_aoi draw or define an AOI as sf, sp or matrix object for the running session that can be used by all query functions.

view_aoi display the session AOI in an interactive mapview/leaflet map viewer.

get_aoi get the session AOI you have defined or drawn before as sf, sp or matrix object.

Further details are included in the README file, which is constantly updated on https://github.com/16EAGLE/getSpatialData.

See also