Get product names

get_products obtains the names of all products available through getSpatialData. The returned names can be used with the get_records functions for querying a specific product.

get_products(product_groups = "all", grouped = FALSE, update_online = FALSE)

getSentinel_products()

getLandsat_products(..., update_online = TRUE)

getMODIS_products(..., update_online = TRUE)

getSRTM_products(..., update_online = TRUE)

Arguments

product_groups

character, either "all" or a combination of product names. Supported names are "sentinel", "landsat", "modis", "srtm".

grouped

logical, whether to return a list of names grouped by platform/sensor (TRUE) or not.

update_online

logical, whether to update the internal product list online (default, requires login) or not.

...

deprecated arguments.

Value

A character vector

Note

If update_online = TRUE, you must be logged in at the services required for your request to use this function. See the examples and login for details.

GNSS

GNSS products (such as "sentinel-1_gnss") retrieved from the dual-frequency GPS recievers mounted on Sentinel-1, -2, and -3 represent a special type of product, as they are AOI-independent and thus only referenced by mission time. GNSS data originally have been used only to precisely calculate the satellites' orbits, but then have been released to the scientific public due to their potential scientifc uses (for details, see https://earth.esa.int/web/sentinel/missions/sentinel-3/news/-/article/new-gnss-l1b-rinex-data-release-for-sentinel-1-2-and-3 and https://earth.esa.int/documents/247904/351187/GMES_Sentinels_POD_Service_File_Format_Specification).

See also

Author

Jakob Schwalb-Willmann

Examples

if (FALSE) {
library(getSpatialData)

# login
login_USGS(username = "your_username")

# get all available products as one vector
products <- get_products()

# or grouped
products <- get_products(grouped = T)
products$sentinel
products$landsat
products$modis

# or get products by a specific service
products <- get_products("sentinel")
products <- get_products("landsat")
products <- get_products("modis")

}