Returns TRUE for records that are of product group 'Sentinel' or the referred sub-selection

These functions check which records are records of the referred product group, product or sub-selection.

is.sentinel(records)

is.sentinel1(records)

is.sentinel1_iw_slc(records)

is.sentinel1_iw_grdh(records)

is.sentinel1_iw_raw(records)

is.sentinel1_iw_ocn(records)

is.sentinel1_level0(records)

is.sentinel1_level1(records)

is.sentinel1_level2(records)

is.sentinel2(records)

is.sentinel2_L1C(records)

is.sentinel2_L2A(records)

is.sentinel2_S2A(records)

is.sentinel2_S2B(records)

is.sentinel3(records)

is.sentinel3_S3A(records)

is.sentinel3_S3B(records)

is.sentinel3_S3C(records)

is.sentinel3_S3D(records)

is.sentinel3_synergy(records)

is.sentinel3_slstr(records)

is.sentinel3_sral(records)

is.sentinel3_olci(records)

is.sentinel5(records)

is.sentinel5p(records)

Arguments

records
data.frame, one or multiple records (each represented by one row), as it is returned by get_records.

Value

logical vector, same length as number of rows in records.

Details

is.sentinel returns TRUE for records that are of product group 'Sentinel'.

is.sentinel1 returns TRUE for records that are of product 'Sentinel-1'.

is.sentinel1_iw_slc returns TRUE for records that are of product 'Sentinel-1' IW SLC (Sentinel-1 Interferometric Wideswath Single Look Complex).

is.sentinel1_iw_grdh returns TRUE for records that are of product 'Sentinel-1' IW GRDH (Interferometric Wideswath Ground Range Detected).

is.sentinel1_iw_raw returns TRUE for records that are of product 'Sentinel-1' IW RAW (Interferometric Wideswath RAW).

is.sentinel1_iw_ocn returns TRUE for records that are of product 'Sentinel-1' IW OCN (Interferometric Wideswath OCN).

is.sentinel1_level0 returns TRUE for records that are of product 'Sentinel-1' level 0.

is.sentinel1_level1 returns TRUE for records that are of product 'Sentinel-1' level 1.

is.sentinel1_level2 returns TRUE for records that are of product 'Sentinel-1' level 2.

is.sentinel2 returns TRUE for records that are of product 'Sentinel-2'.

is.sentinel2_L1C returns TRUE for records that are 'Sentinel-2' Level-1C.

is.sentinel2_L2A returns TRUE for records that are 'Sentinel-2' Level-2A.

is.sentinel2_S2A returns TRUE for records that are of product 'Sentinel-2' and platform S2A.

is.sentinel2_S2B returns TRUE for records that are of product 'Sentinel-2' and platform S2B.

is.sentinel3 returns TRUE for records that are of product 'Sentinel-3'.

is.sentinel3_S3A returns TRUE for records that are of product 'Sentinel-3' and platform S3A.

is.sentinel3_S3B returns TRUE for records that are of product 'Sentinel-3' and platform S3B.

is.sentinel3_S3C returns TRUE for records that are of product 'Sentinel-3' and platform S3C.

is.sentinel3_S3D returns TRUE for records that are of product 'Sentinel-3' and platform S3D.

is.sentinel3_synergy returns TRUE for records that 'Sentinel-3' SYNERGY records.

is.sentinel3_slstr returns TRUE for records that are 'Sentinel-3' SLSTR records.

is.sentinel3_sral returns TRUE for records that are 'Sentinel-3' SRAL records.

is.sentinel3_olci returns TRUE for records that are 'Sentinel-3' OLCI records.

is.sentinel5 and is.sentinel5p return TRUE for records that are 'Sentinel-5' or 'Sentinel-5P'.

Author

Henrik Fisser, 2020