Prepare Sentinel data

prepSentinel makes downloaded Sentinel datasets ready-to-use by automatically extracting, sorting and converting the relevant contents of the datasets to a user-defined format.

prepSentinel(dir_zip, dir_out = NULL, format = "vrt",
  select.tiles = "all", overwrite = F, verbose = T)

Arguments

dir_zip
character, path to zip file.

dir_out
character, full path to target directory. Optional. If not set, getSentinel_data uses the directory to the getSpatialData archive folder. Use set_archive to once define a getSpatialData archive folder.

format
character, output format of the raster datasets (indicated by the file extension, e.g. "tiff" for format GTiff). Default is "vrt".

select.tiles
character, selecting one or multuple specific tiles to prepare. Older datasets archived by ESA contain multiple tiles. If select.tiles is set to one or multiple specific tiles contained within the selected datasets (e.g. "T32TNT"), only these tiles will be extracted and prepared. By default, all tiles are extracted and prepared. Ignored for datasets that contain only a single tile.

overwrite
logical, whether to overwrite existing files or not.

verbose
logical, if TRUE, details on the function's progress will be visibile on the console. Default is TRUE.

Value

Character vector of paths to the prepared files.

See also