View or plot previews of records

view_previews and plot_previews load and display georeferenced previews acquired using get_previews.

view_previews(
  records,
  show_aoi = TRUE,
  aoi_colour = "deepskyblue",
  maxpixels = 1e+05,
  verbose = TRUE
)

plot_previews(
  records,
  show_aoi = TRUE,
  aoi_colour = "deepskyblue",
  separate = FALSE,
  maxcol = 3,
  maxpixels = 1e+05,
  verbose = TRUE
)

Arguments

records

records data frame, containing one or multiple records (each represented by one row), as returned by get_records

show_aoi

logical, whether the session AOI defined with set_aoi should be drawn on the map or plot(s) (TRUE) or not.

aoi_colour

character, colour of the AOI. Ignored, if show_aoi = FALSE.

maxpixels

numeric, maximum number of pixels to be displayed

verbose

logical, whether to display details on the function's progress or output on the console.

separate

logical, whether previews should be displayed separately on a cowplot grid (TRUE) or as a mosaic in a single plot (FALSE, default).

maxcol

numeric, maximum number of columns that the plot grid should have.

Value

A mapview/ggplot object

Details

view_previews renders previews as RGB images on an interactive map using mapview and leaflet.

plot_previews plots previews to a graphics device using ggplot2.

Author

Jakob Schwalb-Willmann